some documentation and correct the error code for AbtractControlUnit phase(now are used negative value for error) added the class for chaos circular buffer

+++ chaos/common/message/DeviceMessageChannel.cpp
@@ -88,7 +88,7 @@ int DeviceMessageChannel::setAttributeValue(CDataWrapper& attributesValues, uint
 }
 
     //------------------------------------
-int DeviceMessageChannel::setScheduleDelay(int64_t scheduledDealy, uint32_t millisecToWait){
+int DeviceMessageChannel::setScheduleDelay(uint32_t scheduledDealy, uint32_t millisecToWait){
     int err = 0;
     CDataWrapper startDeviceParam;
     startDeviceParam.addStringValue(DatasetDefinitionkey::CS_CM_DATASET_DEVICE_ID, deviceNetworkAddress->deviceID);
